content/media/video/test/test_async_abort_events.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478299
-->
<head>
  <title>Test for Bug 478299</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478299">Mozilla Bug 478299</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 478299 **/

// This tests that we bail out of a load if we do another load in an abort handler.
// See: https://bugzilla.mozilla.org/show_bug.cgi?id=465458#c4


var v = null;
var gAbortCount = 0;
var gErrorCount = 0;
var gEmptiedCount = 0;
var gLoadedDataCount = 0;

SimpleTest.waitForExplicitFinish();

function errorListener() {
  gErrorCount++;
}

function emptiedListener() {
  gEmptiedCount++;
}

function loadedDataListener() {
  gLoadedDataCount++;
  is(gLoadedDataCount, 1, "Should only get 1 load");
  is(gAbortCount, 1, "Should only get 1 abort");
  is(gEmptiedCount, 1, "Should only get 1 emptied");
  is(gErrorCount, 0, "Shouldn't get any errors");
  SimpleTest.finish();
}

function abortListener() {
  gAbortCount++;
  if (gAbortCount == 1) {
    v.src = '320x240.ogv';
    v.load();
  }
}

v = document.createElement('video');
v.addEventListener('error', errorListener, false);
v.addEventListener('emptied', emptiedListener, false);
v.addEventListener('loadeddata', loadedDataListener, false);
v.addEventListener('abort', abortListener, false);

is(v.networkState, HTMLMediaElement.NETWORK_EMPTY, "NW State invalid.");
v.load();
ok(v.networkState > HTMLMediaElement.NETWORK_EMPTY, "NW State invalid.");

// Cause another load. This will abort the previous load, causing 'abort' event to fire.
// That will trigger the abort handler.
// Abort handler adds 'src' attribute, and starts a new load. This aborts the previous load,
// so it won't fire emptied for the first abort. The second abort then causes 'abort' and
// 'emptied' events. The load should then succeed, causing 'loadeddata' event.
v.load();


</script>
</pre>
</body>
</html>
